# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A web API for live RC timing data.",
        "title": "RC Timing API",
        "contact": {
            "name": "FaintLocket424",
            "url": "https://github.com/FaintLocket424/rc-timing-api"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/event/practice/heat/{id}/round/{round_id}": {
            "get": {
                "description": "Retrieves results for a specific practice heat and round using the target timing software.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "practice"
                ],
                "summary": "Get practice heat results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target Timing URL (URL-encoded)",
                        "name": "url",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Timing software type (default: bbk)",
                        "name": "software",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Heat Number",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Round Number",
                        "name": "round_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_FaintLocket424_rc-timing-api_internal_models.HeatResultDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid Heat or Round ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Scraper or Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_FaintLocket424_rc-timing-api_internal_models.DriverResultDTO": {
            "type": "object",
            "properties": {
                "best_lap": {
                    "$ref": "#/definitions/time.Duration"
                },
                "car_number": {
                    "type": "integer"
                },
                "driver_name": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "race_time": {
                    "$ref": "#/definitions/github_com_FaintLocket424_rc-timing-api_internal_models.RaceTimeDTO"
                }
            }
        },
        "github_com_FaintLocket424_rc-timing-api_internal_models.HeatResultDTO": {
            "type": "object",
            "properties": {
                "best_lap": {
                    "$ref": "#/definitions/time.Duration"
                },
                "class": {
                    "type": "string"
                },
                "class_best_lap": {
                    "$ref": "#/definitions/time.Duration"
                },
                "class_best_time": {
                    "$ref": "#/definitions/github_com_FaintLocket424_rc-timing-api_internal_models.RaceTimeDTO"
                },
                "heat_number": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_FaintLocket424_rc-timing-api_internal_models.DriverResultDTO"
                    }
                },
                "round_number": {
                    "type": "integer"
                }
            }
        },
        "github_com_FaintLocket424_rc-timing-api_internal_models.RaceTimeDTO": {
            "type": "object",
            "properties": {
                "laps": {
                    "type": "integer"
                },
                "race_time": {
                    "$ref": "#/definitions/time.Duration"
                }
            }
        },
        "time.Duration": {
            "type": "integer",
            "format": "int64",
            "enum": [
                -9223372036854775808,
                9223372036854775807,
                1,
                1000,
                1000000,
                1000000000,
                60000000000,
                3600000000000
            ],
            "x-enum-varnames": [
                "minDuration",
                "maxDuration",
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second",
                "Minute",
                "Hour"
            ]
        }
    }
}